requirements.txt
flask
cognitive_faces
glob2